metrics-agent/requirements.txt
grpcio
grpcio-tools
psutil
pymongo
python-dotenv
requests
